refactor SessionLog
